Only give instructions to install as npm/yarn dependency
I don't know when it would make sense to have this as a global dependency. I feel like for the most part as a developer consuming your library (which is the reason I'm here) I would just want instructions to include this in my project. Adding as a global dependency doesn't offer much value to me and I feel like might cause confusion when this works for someone locally but doesn't work for another developer. Similarly I feel like cloning promotes an anti pattern that could cause some developers to be tempted to start their own fork of this library and miss out on important updates.